long long long long time

+++ assignment2/minDotProduct/main.c
@@ -2,38 +2,38 @@
 #include <stdlib.h>
 
 int compareInts(const void *a, const void *b) {
-  const int *ia = (const int *)a;
-  const int *ib = (const int *)b;
+  const long long *ia = (const long long *)a;
+  const long long *ib = (const long long *)b;
   return *ia - *ib;
 }
 
 int reverseInts(const void *a, const void *b) {
-  const int *ia = (const int *)a;
-  const int *ib = (const int *)b;
+  const long long *ia = (const long long *)a;
+  const long long *ib = (const long long *)b;
   return *ib - *ia;
 }
 
 int main() {
-  int n, result, i;
+  long long n, result, i;
   result = 0;
 
-  scanf("%d\n", &n);
-  int listOne[n], listTwo[n];
+  scanf("%lld\n", &n);
+  long long listOne[n], listTwo[n];
 
   for (i = 0; i < n; i++) {
-    scanf("%d", &listOne[i]);
+    scanf("%lld", &listOne[i]);
   }
 
   for (i = 0; i < n; i++) {
-    scanf("%d", &listTwo[i]);
+    scanf("%lld", &listTwo[i]);
   }
 
-  qsort(listOne, n, sizeof(int), compareInts);
-  qsort(listTwo, n, sizeof(int), reverseInts);
+  qsort(listOne, n, sizeof(long long), compareInts);
+  qsort(listTwo, n, sizeof(long long), reverseInts);
 
   for (i = 0; i < n; i++) {
     result += listOne[i] * listTwo[i];
   }
 
-  printf("%d\n", result);
+  printf("%lld\n", result);
 }
